changin the data type to long

+++ modules/sharing-registry/sharing-registry-server/src/main/java/org/apache/airavata/sharing/registry/db/repositories/EntityRepository.java
@@ -146,7 +146,7 @@ public class EntityRepository extends AbstractRepository<Entity, EntityEntity, E
             entity.setBinaryData((byte[]) (rs[7]));
             entity.setFullText((String) (rs[8]));
             entity.setOriginalEntityCreationTime((long)(rs[9]));
-            entity.setSharedCount((int) rs[10]);
+            entity.setSharedCount((long) rs[10]);
             entity.setCreatedTime((long) (rs[11]));
             entity.setUpdatedTime((long) (rs[12]));
